Evening upload; Setting Update in progress; v2.5.0
Still work in progress & need to be redesigned (it's useless and heavy to change the destination AND the name of the output folder. A filedialog to choose the outputfolder is enough).

Since this commit, the script is a bit buggy but works anyway. Need to be optimized for the future.